# optimize number serializers