Benchmark

For 1 Million serialize/deserialize:

with u32
old version: 416.561ms
new version: 296.325ms
gain:  28.8%

with u64
old version: 515.476ms
new version: 305.632ms
gain:  40.7%

with f64
old version: 492.263ms
new version: 321.706ms
gain:  34.7%